I kinda made a few changes.
I added a new class called TaskList which handles parsing and shit.
I changed how we parse commands with it.  I felt like SystemController
was doing too much of the work so I split that out of that class.  It
seems a lot easier to handle, fix, test, you name it.  I didn't change
how any of the commands are actually executed, just how they are parsed
and read in.  I added all the explanations in the files so take a gander
at those bad boys.  Have a good one fellas!